+++ Project 06 - FindMyLocation/Where/ViewController.swift
@@ -9,10 +9,13 @@
 import UIKit
 import CoreLocation
 
-class ViewController: UIViewController, CLLocationManagerDelegate {
+class ViewController: UIViewController {
 
     @IBOutlet weak var locationLabel: UILabel!
     
+    // 强制自动解包，可以赋值为nil，为nil后再调用会报错
+    // 建议定义为：
+    // var locationManager: CLLocationManager
     var locationManager: CLLocationManager!
     
     override func viewDidLoad() {
@@ -38,49 +41,6 @@ class ViewController: UIViewController, CLLocationManagerDelegate {
         super.didReceiveMemoryWarning()
         // Dispose of any resources that can be recreated.
     }
-    
-    func locationManager(_ manager: CLLocationManager, didFailWithError error: Error) {
-        
-      self.locationLabel.text = "Error while updating location " + error.localizedDescription
-        
-    }
-    
-    func locationManager(_ manager: CLLocationManager, didUpdateLocations locations: [CLLocation]) {
-        CLGeocoder().reverseGeocodeLocation(manager.location!, completionHandler: {(placemarks, error)->Void in
-            
-            if (error != nil) {
-               self.locationLabel.text = "Reverse geocoder failed with error" + error!.localizedDescription
-                return
-            }
-            
-            if placemarks!.count > 0 {
-                let pm = placemarks![0] 
-                self.displayLocationInfo(pm)
-            } else {
-               self.locationLabel.text = "Problem with the data received from geocoder"
-            }
-        })
-    }
-    
-    func displayLocationInfo(_ placemark: CLPlacemark?) {
-        if let containsPlacemark = placemark {
-            //stop updating location to save battery life
-            locationManager.stopUpdatingLocation()
-            
-            let locality = (containsPlacemark.locality != nil) ? containsPlacemark.locality : ""
-            let postalCode = (containsPlacemark.postalCode != nil) ? containsPlacemark.postalCode : ""
-            let administrativeArea = (containsPlacemark.administrativeArea != nil) ? containsPlacemark.administrativeArea : ""
-            let country = (containsPlacemark.country != nil) ? containsPlacemark.country : ""
-            
-            self.locationLabel.text = postalCode! + " " + locality!
-            
-            self.locationLabel.text?.append("\n" + administrativeArea! + ", " + country!)
-        }
-        
-    }
-
-    
-    
 
 }
